Made TrackEvaluationContainer an empty base class Put everything in TrackEvaluationContainerv1 updated TrackEvaluation module accordingly Unfortunately, this is broken. I cannot read the files any more
