+++ simulation/g4simulation/g4eval/TrackEvaluationContainerv1LinkDef.h
@@ -1,0 +1,5 @@
+#ifdef __CINT__
+
+#pragma link C++ class TrackEvaluationContainerv1+;
+
+#endif /* __CINT__ */
